more 10.5 support
git-svn-id: https://svn.code.sf.net/p/osirix/code/osirix@2823 4e9e6e3d-4551-47eb-9dbe-a6127f9400a1

+++ ROI.m
@@ -4573,7 +4573,7 @@ void gl_round_box(int mode, float minx, float miny, float maxx, float maxy, floa
 	layerPixelSpacingY = y;
 }
 
-int sortPointArrayAlongX(id point1, id point2, void *context)
+NSInteger sortPointArrayAlongX(id point1, id point2, void *context)
 {
     float x1 = (float)[point1 pointValue].x;
     float x2 = (float)[point2 pointValue].x;
